+++ tests/api_resources/test_namespaces.py
@@ -172,6 +172,8 @@ class TestNamespaces:
     @parametrize
     def test_method_explain_query_with_all_params(self, client: Turbopuffer) -> None:
         namespace = client.namespace("namespace").explain_query(
+            aggregate_by={"foo": ("Sum", "bar")},
+            compute_attributes={"foo": ("Highlight", "bar")},
             consistency={"level": "strong"},
             distance_metric="cosine_distance",
             exclude_attributes=["string"],
@@ -294,6 +296,8 @@ class TestNamespaces:
             namespace="namespace",
             queries=[
                 {
+                    "aggregate_by": {"foo": "bar"},
+                    "compute_attributes": {"foo": ("Highlight", "bar")},
                     "distance_metric": "cosine_distance",
                     "exclude_attributes": ["string"],
                     "include_attributes": True,
@@ -355,6 +359,8 @@ class TestNamespaces:
     @parametrize
     def test_method_query_with_all_params(self, client: Turbopuffer) -> None:
         namespace = client.namespace("namespace").query(
+            aggregate_by={"foo": ("Sum", "bar")},
+            compute_attributes={"foo": ("Highlight", "bar")},
             consistency={"level": "strong"},
             distance_metric="cosine_distance",
             exclude_attributes=["string"],
@@ -793,6 +799,8 @@ class TestAsyncNamespaces:
     @parametrize
     async def test_method_explain_query_with_all_params(self, async_client: AsyncTurbopuffer) -> None:
         namespace = await async_client.namespace("namespace").explain_query(
+            aggregate_by={"foo": ("Sum", "bar")},
+            compute_attributes={"foo": ("Highlight", "bar")},
             consistency={"level": "strong"},
             distance_metric="cosine_distance",
             exclude_attributes=["string"],
@@ -918,6 +926,8 @@ class TestAsyncNamespaces:
             namespace="namespace",
             queries=[
                 {
+                    "aggregate_by": {"foo": "bar"},
+                    "compute_attributes": {"foo": ("Highlight", "bar")},
                     "distance_metric": "cosine_distance",
                     "exclude_attributes": ["string"],
                     "include_attributes": True,
@@ -979,6 +989,8 @@ class TestAsyncNamespaces:
     @parametrize
     async def test_method_query_with_all_params(self, async_client: AsyncTurbopuffer) -> None:
         namespace = await async_client.namespace("namespace").query(
+            aggregate_by={"foo": ("Sum", "bar")},
+            compute_attributes={"foo": ("Highlight", "bar")},
             consistency={"level": "strong"},
             distance_metric="cosine_distance",
             exclude_attributes=["string"],